# xds: refactor inbound listener generation