… pilot/pkg/networking/core/v1alpha3/gateway_test.go
… pilot/pkg/networking/core/v1alpha3/listener.go
… pilot/pkg/networking/core/v1alpha3/listener_builder.go
… pilot/pkg/networking/core/v1alpha3/listener_builder_test.go
… pilot/pkg/networking/core/v1alpha3/networkfilter_test.go
… pilot/pkg/networking/core/v1alpha3/sidecar_simulation_test.go
… pilot/pkg/networking/core/v1alpha3/tracing.go
… pilot/pkg/networking/util/util.go
… pilot/pkg/simulation/traffic.go
… pilot/test/xdstest/test.go