Currently, the listener generation code has decayed substantially as
features have been bolted on over the years. The biggest issue is that
we build `FilterChain`s, then build them into real `Listener`s, but then later convert them back to just
`FilterChain`s. This is not only super complex, but also inefficient --
benchmarks below show about 5-15% improvement just by not doing this.
Further, it limits our ability to make changes, such as taking advantage
of the new `FilterChainMatcher` in Envoy.

Some other issues are that we have a bunch of poor coding practices,
like fake instances of structs (ServiceInstance), mutable state
(`listenerMap`), etc. We have a bunch of extremely complex config to
build up a variety of filters then try to arrange them appropriately
later, rather than just building them correctly in the first place.

This PR refactors all *inbound* listeners, following
https://github.com/istio/istio/pull/38595, to a completely new code
path. We have pretty solid test coverage here, and I have done manual
XDS comparison of before/after to ensure no behavioral changes.

The new code behaves a bit differently than before. First, we build up a
'plan' for the listener. This allows us to handle conflicts, etc, in
simple code rather than interweaved with the xDS generation. Then we use
the plan to build up the listeners/filter chains.

Benchmarks:
```
name                                             old time/op        new time/op        delta
ListenerGeneration/gateways-6                          18.8ms ± 8%        18.6ms ± 4%     ~     (p=0.537 n=6+5)
ListenerGeneration/gateways-shared-6                   39.8µs ± 8%        39.0µs ± 4%     ~     (p=0.485 n=6+6)
ListenerGeneration/empty-6                             1.43ms ± 4%        1.36ms ± 6%   -4.61%  (p=0.041 n=6+6)
ListenerGeneration/tls-6                               1.40ms ±15%        1.33ms ± 4%     ~     (p=0.240 n=6+6)
ListenerGeneration/telemetry-6                         1.58ms ±10%        1.66ms ± 6%     ~     (p=0.240 n=6+6)
ListenerGeneration/telemetry-api-6                     1.57ms ± 8%        1.47ms ± 7%   -6.13%  (p=0.026 n=6+6)
ListenerGeneration/virtualservice-6                     117µs ± 3%         113µs ± 4%   -3.58%  (p=0.041 n=6+6)
ListenerGeneration/authorizationpolicy-6               1.28ms ± 4%        1.18ms ± 7%   -7.71%  (p=0.015 n=6+6)
ListenerGeneration/peerauthentication-6                89.6µs ± 6%        73.9µs ± 1%  -17.56%  (p=0.004 n=6+5)
ListenerGeneration/knative-gateway-6                   35.6µs ± 5%        34.9µs ± 2%     ~     (p=0.247 n=6+5)
ListenerGeneration/serviceentry-workloadentry-6        1.41ms ± 1%        1.36ms ± 7%     ~     (p=0.093 n=6+6)

name                                             old kb/msg         new kb/msg         delta
ListenerGeneration/gateways-6                           2.11k ± 0%         2.11k ± 0%     ~     (all equal)
ListenerGeneration/gateways-shared-6                     4.02 ± 0%          4.02 ± 0%     ~     (all equal)
ListenerGeneration/empty-6                               18.4 ± 0%          18.5 ± 0%   +0.49%  (p=0.002 n=6+6)
ListenerGeneration/tls-6                                 12.8 ± 0%          12.6 ± 0%   -1.02%  (p=0.002 n=6+6)
ListenerGeneration/telemetry-6                           31.6 ± 0%          31.7 ± 0%   +0.32%  (p=0.002 n=6+6)
ListenerGeneration/telemetry-api-6                       23.9 ± 0%          24.0 ± 0%   +0.42%  (p=0.002 n=6+6)
ListenerGeneration/virtualservice-6                      7.43 ± 0%          7.32 ± 0%   -1.60%  (p=0.002 n=6+6)
ListenerGeneration/authorizationpolicy-6                 65.2 ± 0%          38.9 ± 0%  -40.34%  (p=0.002 n=6+6)
ListenerGeneration/peerauthentication-6                  6.55 ± 0%          6.41 ± 0%   -2.11%  (p=0.002 n=6+6)
ListenerGeneration/knative-gateway-6                     3.56 ± 0%          3.56 ± 0%     ~     (all equal)
ListenerGeneration/serviceentry-workloadentry-6          18.4 ± 0%          18.5 ± 0%   +0.49%  (p=0.002 n=6+6)

name                                             old resources/msg  new resources/msg  delta
ListenerGeneration/gateways-6                            2.00 ± 0%          2.00 ± 0%     ~     (all equal)
ListenerGeneration/gateways-shared-6                     2.00 ± 0%          2.00 ± 0%     ~     (all equal)
ListenerGeneration/empty-6                               6.00 ± 0%          6.00 ± 0%     ~     (all equal)
ListenerGeneration/tls-6                                 6.00 ± 0%          6.00 ± 0%     ~     (all equal)
ListenerGeneration/telemetry-6                           6.00 ± 0%          6.00 ± 0%     ~     (all equal)
ListenerGeneration/telemetry-api-6                       6.00 ± 0%          6.00 ± 0%     ~     (all equal)
ListenerGeneration/virtualservice-6                      3.00 ± 0%          3.00 ± 0%     ~     (all equal)
ListenerGeneration/authorizationpolicy-6                 4.00 ± 0%          4.00 ± 0%     ~     (all equal)
ListenerGeneration/peerauthentication-6                  4.00 ± 0%          4.00 ± 0%     ~     (all equal)
ListenerGeneration/knative-gateway-6                     2.00 ± 0%          2.00 ± 0%     ~     (all equal)
ListenerGeneration/serviceentry-workloadentry-6          6.00 ± 0%          6.00 ± 0%     ~     (all equal)

name                                             old alloc/op       new alloc/op       delta
ListenerGeneration/gateways-6                          10.4MB ± 0%        10.4MB ± 0%   +0.15%  (p=0.002 n=6+6)
ListenerGeneration/gateways-shared-6                   23.0kB ± 0%        23.0kB ± 0%   -0.14%  (p=0.002 n=6+6)
ListenerGeneration/empty-6                              787kB ± 0%         757kB ± 0%   -3.79%  (p=0.004 n=6+5)
ListenerGeneration/tls-6                                742kB ± 0%         729kB ± 0%   -1.79%  (p=0.010 n=6+4)
ListenerGeneration/telemetry-6                          891kB ± 0%         861kB ± 0%   -3.35%  (p=0.004 n=6+5)
ListenerGeneration/telemetry-api-6                      816kB ± 0%         781kB ± 0%   -4.26%  (p=0.002 n=6+6)
ListenerGeneration/virtualservice-6                    81.3kB ± 0%        74.9kB ± 0%   -7.84%  (p=0.008 n=5+5)
ListenerGeneration/authorizationpolicy-6                609kB ± 0%         515kB ± 0%  -15.38%  (p=0.002 n=6+6)
ListenerGeneration/peerauthentication-6                47.4kB ± 0%        40.1kB ± 0%  -15.44%  (p=0.026 n=5+6)
ListenerGeneration/knative-gateway-6                   21.4kB ± 0%        21.4kB ± 0%   -0.15%  (p=0.002 n=6+6)
ListenerGeneration/serviceentry-workloadentry-6         787kB ± 0%         757kB ± 0%   -3.80%  (p=0.002 n=6+6)

name                                             old allocs/op      new allocs/op      delta
ListenerGeneration/gateways-6                            113k ± 0%          113k ± 0%     ~     (p=0.892 n=6+6)
ListenerGeneration/gateways-shared-6                      290 ± 0%           290 ± 0%     ~     (all equal)
ListenerGeneration/empty-6                              14.7k ± 0%         14.3k ± 0%   -2.36%  (p=0.002 n=6+6)
ListenerGeneration/tls-6                                14.2k ± 0%         14.0k ± 0%   -1.39%  (p=0.002 n=6+6)
ListenerGeneration/telemetry-6                          16.0k ± 0%         15.6k ± 0%   -2.16%  (p=0.002 n=6+6)
ListenerGeneration/telemetry-api-6                      16.1k ± 0%         15.7k ± 0%   -2.51%  (p=0.002 n=6+6)
ListenerGeneration/virtualservice-6                       968 ± 0%           894 ± 0%   -7.64%  (p=0.002 n=6+6)
ListenerGeneration/authorizationpolicy-6                10.5k ± 0%         10.3k ± 0%   -1.73%  (p=0.002 n=6+6)
ListenerGeneration/peerauthentication-6                   602 ± 0%           509 ± 0%  -15.45%  (p=0.002 n=6+6)
ListenerGeneration/knative-gateway-6                      274 ± 0%           274 ± 0%     ~     (all equal)
ListenerGeneration/serviceentry-workloadentry-6         14.7k ± 0%         14.3k ± 0%   -2.36%  (p=0.002 n=6+6)
```

**Please provide a description of this PR:**