@istio-testing 2022-04-28
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
integ-telemetry_istio | 12ff33f47590abc68360488e2d5fc301dbd2a556 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/38652/integ-telemetry_istio/1519752028652310528) | true | `/test integ-telemetry_istio`
integ-security_istio | 12ff33f47590abc68360488e2d5fc301dbd2a556 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/38652/integ-security_istio/1519752031869341696) | true | `/test integ-security_istio`
integ-telemetry-istiodremote_istio | 12ff33f47590abc68360488e2d5fc301dbd2a556 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/38652/integ-telemetry-istiodremote_istio/1519752028727808000) | true | `/test integ-telemetry-istiodremote_istio`
integ-telemetry-mc_istio | 12ff33f47590abc68360488e2d5fc301dbd2a556 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/38652/integ-telemetry-mc_istio/1519752028685864960) | true | `/test integ-telemetry-mc_istio`
integ-security-istiodremote_istio | 12ff33f47590abc68360488e2d5fc301dbd2a556 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/38652/integ-security-istiodremote_istio/1519752033547063296) | true | `/test integ-security-istiodremote_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->